Change footnote order
Previously, the footer (if footnote definitions were found) showed all footnote
definitions, in definition order.  Even footnote definitions with the same
identifier were transformed to multiple definitions.

remarkjs/remark-rehype#11 shows that definition order does not make sense if
inline footnotes (`[^Some text]`) are used in combination with footnote
references (`[^id]`) and footnote definitions (`[^id]: Some text`).

Investigating this, I came to the conclusion that it also does not make sense to
output all definitions, whether they are used (or even duplicate) or not.

Instead, it makes sense to track the order in which footnotes are used (first),
and only output a definition for the used inline footnotes / footnote
references, in the order they were used.  This is in line with how link
definitions are used as well.

Closes remarkjs/remark-rehype#11.
Closes GH-32.