Fix to always add `id` to footnote heading
Previusly, since 1dc8ab3, it was possible to customize the properties
generated on the `h2` element in the footnote section.
That commit made it possible to remove the `id` attribute.
However, other steps in the process require that that `id` attribute,
with the expected value, exists.

Due to that, broken, inaccessible HTML could be generated.
This is now fixed.

Related-to: GH-64.